Clean up logging code to reduce code duplication.
