+++ src/libraries/Highbyte.DotNet6502/Logging/Console/DotNet6502ConsoleLogger.cs
@@ -4,98 +4,19 @@ using Microsoft.Extensions.ObjectPool;
 
 namespace Highbyte.DotNet6502.Logging.Console;
 
-public class DotNet6502ConsoleLogger : ILogger
+public class DotNet6502ConsoleLogger : DotNet6502LoggerBase
 {
     private readonly Func<DotNet6502ConsoleLoggerConfiguration> _getCurrentConfig;
-    private readonly ObjectPool<StringBuilder> _stringBuilderPool;
-    private readonly string _categoryName;
 
     public DotNet6502ConsoleLogger(
         ObjectPool<StringBuilder> stringBuilderPool,
         string categoryName,
-        Func<DotNet6502ConsoleLoggerConfiguration> getCurrentConfig)
+        Func<DotNet6502ConsoleLoggerConfiguration> getCurrentConfig) : base(stringBuilderPool, categoryName)
     {
-        _stringBuilderPool = stringBuilderPool;
-        _categoryName = categoryName;
         _getCurrentConfig = getCurrentConfig;
     }
 
-    public IDisposable BeginScope<TState>(TState state) => default!;
-
-    public bool IsEnabled(LogLevel logLevel) => logLevel >= _getCurrentConfig().LogLevel;
-
-    public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
-    {
-        if (!IsEnabled(logLevel))
-            return;
-
-        if (formatter is null)
-            throw new ArgumentNullException(nameof(formatter));
-
-        var message = formatter(state, exception);
-
-        if (string.IsNullOrEmpty(message) && exception is null)
-            return;
-
-        WriteMessage(logLevel, eventId.Id, message, exception);
-    }
-
-    private void WriteMessage(LogLevel logLevel, int eventId, string message, Exception ex)
-    {
-        var builder = _stringBuilderPool.Get();
-        try
-        {
-            var time = DateTime.Now.ToString("HH:mm:ss.fff");
-            //var threadId = Environment.CurrentManagedThreadId;
-
-            builder
-                .Append(time)
-                .Append(" [")
-                .Append(GetLogLevelString(logLevel))
-                .Append("] ")
-                .Append(_categoryName)
-                .Append(' ')
-                //.Append("(")
-                //.Append(eventId)
-                //.Append(") ")
-                //.Append(" (")
-                //.Append(threadId)
-                //.Append(") ")
-                .AppendLine(message);
-
-            if (ex is { })
-            {
-                builder.Append("    ");
-                builder.AppendLine(ex.ToString());
-            }
-
-            var logMsg = builder.ToString();
-
-            System.Console.WriteLine(logMsg);
-        }
-        finally
-        {
-            _stringBuilderPool.Return(builder);
-        }
-
-        static string GetLogLevelString(LogLevel logLevel) => logLevel switch
-        {
-            LogLevel.Trace => "trce",
-            LogLevel.Debug => "dbug",
-            LogLevel.Information => "info",
-            LogLevel.Warning => "warn",
-            LogLevel.Error => "fail",
-            LogLevel.Critical => "crit",
-            _ => throw new ArgumentOutOfRangeException(nameof(logLevel)),
-        };
-    }
-}
-
-internal class NullDisposable : IDisposable
-{
-    public static IDisposable Instance { get; } = new NullDisposable();
-
-    public void Dispose()
-    {
-    }
+    public override IDisposable BeginScope<TState>(TState state) => default!;
+    public override bool IsEnabled(LogLevel logLevel) => logLevel >= _getCurrentConfig().LogLevel;
+    public override void WriteLog(string message) => System.Console.WriteLine(message);
 }
