+++ src/libraries/Highbyte.DotNet6502/Logging/Console/DotNet6502ConsoleLoggerConfiguration.cs
@@ -1,0 +1,11 @@
+using Microsoft.Extensions.Logging;
+
+namespace Highbyte.DotNet6502.Logging.Console;
+
+public class DotNet6502ConsoleLoggerConfiguration
+{
+    public int EventId { get; set; }
+
+    public LogLevel LogLevel { get; set; } = LogLevel.Information;
+
+}
